fix: fix bug in run_upc_detection
An error was raised for monochrome images (not RGB)
See
https://openfoodfacts.sentry.io/issues/4166627511
