+++ tests/unit/utils/test_module.py
@@ -80,7 +80,8 @@ def test_get_image_from_url_decompression_bomb(mocker):
     session_mock = mocker.Mock()
     response_mock = mocker.Mock()
     mocker.patch(
-        "robotoff.utils.Image", **{"open.side_effect": Image.DecompressionBombError()}
+        "robotoff.utils.image.Image",
+        **{"open.side_effect": Image.DecompressionBombError()}
     )
     response_mock.content = generate_image()
     response_mock.ok = True
